requirements.txt
requests
beautifulsoup4
python-dateutil
feedparser
anthropic
markdown
python-dotenv
